feat(observability): add distributed tracing and centralized logging
- add Gateway observability stack with Prometheus, Grafana, and Tempo configs

- enable Micrometer observation/tracing + Kafka listener observation across Gateway, Order, Inventory, Product, and Notification services

- add service-level observation configuration classes and logback-spring setups for consistent structured logging

- update application properties and pom dependencies required for metrics/tracing pipelines

- include startup guide and supporting API/service adjustments made alongside observability rollout

+++ SPRINGBOOT-Microservices/Product-Service/src/main/java/com/Shlok/Product/dto/ProductRequest.java
@@ -1,7 +1,5 @@
 package com.Shlok.Product.dto;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -17,6 +15,6 @@ public class ProductRequest {
     private String name;
     private String description;
     private String skuCode;
-    private BigDecimal price;
+    private Double price;
 
 }
